fix: search for long week day name as it is being used as title of the abbreviation

+++ src/components/Calendar/CalendarWeeks/CalendarWeeks.spec.tsx
@@ -8,7 +8,7 @@ describe("CalendarWeeks", () => {
     render(<CalendarWeeks />);
     weekDaysNames().forEach((weekDay) => {
       expect(
-        screen.getByRole("columnheader", { name: weekDay.short })
+        screen.getByRole("columnheader", { name: weekDay.long })
       ).toBeInTheDocument();
     });
   });
